requirements.txt
flask==2.3.3
openai>=1.50.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
torch>=1.9.0
opencv-python==4.8.1.78
mediapipe==0.10.7
pandas==2.0.3
numpy>=1.21.0 